+ Changed INT to Float and corrected functionality in FE.

+++ gui/src/queries/addWeight.ts
@@ -1,7 +1,7 @@
 import { gql } from "@apollo/client";
 
 export const ADD_WEIGHT = gql`
-  mutation AddWeight($id: String!, $date: String!, $weight: Int!) {
+  mutation AddWeight($id: String!, $date: String!, $weight: Float!) {
     addWeight(id: $id, date: $date, weight: $weight) {
       id
       date
