+++ gui/src/queries/updateWeight.ts
@@ -1,7 +1,7 @@
 import { gql } from "@apollo/client";
 
 export const UPDATE_WEIGHT = gql`
-  mutation UpdateWeight($id: String!, $date: String!, $weight: Int!) {
+  mutation UpdateWeight($id: String!, $date: String!, $weight: Float!) {
     editWeight(id: $id, date: $date, weight: $weight) {
       id
       date
